backend-service0/requirements.txt
pipenv
django
uuid
djangorestframework
djangorestframework-simplejwt
django-cors-headers
